Refactor NewCommandTestUtil
