+++ src/test/java/seedu/address/logic/commands/room/AddRoomsCommandTest.java
@@ -1,10 +1,10 @@
 package seedu.address.logic.commands.room;
 
-import static seedu.address.logic.commands.NewCommandTestUtil.assertCommandFailure;
-import static seedu.address.logic.commands.NewCommandTestUtil.assertCommandSuccess;
 import static seedu.address.logic.commands.room.InitRoomCommand.MESSAGE_NEGATIVE_VALUES_CANNOT_BE_INPUT;
 import static seedu.address.logic.commands.room.InitRoomCommand.MESSAGE_SUCCESS;
 import static seedu.address.logic.commands.room.InitRoomCommand.MESSAGE_ZERO_CANNOT_BE_AN_INPUT;
+import static seedu.address.testutil.command.GeneralCommandTestUtil.assertCommandFailure;
+import static seedu.address.testutil.command.GeneralCommandTestUtil.assertCommandSuccess;
 
 import org.junit.jupiter.api.Test;
