+++ src/test/java/seedu/address/logic/commands/room/FindRoomCommandTest.java
@@ -1,7 +1,7 @@
 package seedu.address.logic.commands.room;
 
-import static seedu.address.logic.commands.NewCommandTestUtil.assertCommandFailure;
-import static seedu.address.logic.commands.NewCommandTestUtil.assertCommandSuccess;
+import static seedu.address.testutil.command.GeneralCommandTestUtil.assertCommandFailure;
+import static seedu.address.testutil.command.GeneralCommandTestUtil.assertCommandSuccess;
 
 import java.util.function.Predicate;
