+++ src/test/java/seedu/address/logic/commands/room/SearchRoomCommandTest.java
@@ -6,9 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_PATIENT_NAME;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_ROOM_NUMBER;
 import static seedu.address.commons.core.Messages.MESSAGE_PATIENT_NO_ROOM;
-import static seedu.address.logic.commands.NewCommandTestUtil.VALID_NAME_BOB;
-import static seedu.address.logic.commands.NewCommandTestUtil.assertCommandFailure;
-import static seedu.address.logic.commands.NewCommandTestUtil.assertCommandSuccess;
 import static seedu.address.logic.commands.room.SearchRoomCommand.MESSAGE_SUCCESS;
 import static seedu.address.testutil.Assert.assertThrows;
 import static seedu.address.testutil.TypicalPatients.BENSON;
@@ -18,6 +15,9 @@ import static seedu.address.testutil.TypicalRooms.ROOM7_PATIENT_ALICE_NO_TASK;
 import static seedu.address.testutil.TypicalRooms.ROOM8_PATIENT_BENSON_NO_TASK;
 import static seedu.address.testutil.TypicalRooms.ROOM_NUMBER_7;
 import static seedu.address.testutil.TypicalRooms.getTypicalRoomList;
+import static seedu.address.testutil.command.GeneralCommandTestUtil.assertCommandFailure;
+import static seedu.address.testutil.command.GeneralCommandTestUtil.assertCommandSuccess;
+import static seedu.address.testutil.command.PatientCommandTestUtil.VALID_NAME_BOB;
 
 import java.util.Arrays;
