+++ src/test/java/seedu/address/logic/commands/task/EditTaskCommandTest.java
@@ -2,13 +2,13 @@ package seedu.address.logic.commands.task;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.address.logic.commands.NewCommandTestUtil.VALID_DESCRIPTION_REMIND_PATIENT;
-import static seedu.address.logic.commands.NewCommandTestUtil.VALID_ROOM_NUMBER_EIGHT;
-import static seedu.address.logic.commands.NewCommandTestUtil.VALID_ROOM_NUMBER_SEVEN;
-import static seedu.address.logic.commands.NewCommandTestUtil.VALID_TASK_INDEX_ONE;
-import static seedu.address.logic.commands.NewCommandTestUtil.VALID_TASK_INDEX_TWO;
 import static seedu.address.logic.commands.task.EditTaskCommand.EditTaskDescriptor;
 import static seedu.address.testutil.Assert.assertThrows;
+import static seedu.address.testutil.command.RoomCommandTestUtil.VALID_ROOM_NUMBER_EIGHT;
+import static seedu.address.testutil.command.RoomCommandTestUtil.VALID_ROOM_NUMBER_SEVEN;
+import static seedu.address.testutil.command.TaskCommandTestUtil.VALID_DESCRIPTION_REMIND_PATIENT;
+import static seedu.address.testutil.command.TaskCommandTestUtil.VALID_TASK_INDEX_ONE;
+import static seedu.address.testutil.command.TaskCommandTestUtil.VALID_TASK_INDEX_TWO;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
